Correct bug in key logic within update_states_db()
There was a previous change from states[i_diff] to db_states[i_diff] when
determining the datestart after which to drop cmd_states.  However this
conflicted with existing code that set the start time of db_states[-1] to 1970
(to force a mismatch).  It was happening that the db_states[-1] datestart got
used for deleting, which wiped the entire table clean.  This was a bad way to
have done it in the first place and the new code finds a cleaner way of getting
the desired result.