Hardwire task_schedule_occ.cfg to use flight Ska h5file
In the previous PR, update_cmd_states was modified to define a var
is_flight that knows if the cmd_states H5 file is the "Flight" version.
If not it uses a different "test" directory on lucky to avoid conflicts
with flight processing.  With this in place we need to have the
OCC --h5file hardwired to /proj/sot/ska/data because this process runs
in Ska test on GRETA.