Cut the bogus SCS107 and manual interrupt at '2014:187:21:21:52.154'
These non-load commands and interrupt were inserted as a convenience
to at least have the pointing correct in cmd_states for propagation
after the 2014:187 commanding problem.  Now that the load_segments
appear correct from iFOT, the interrupt and bogus SCS107 at this time
should not exist in the nonload commands archive.

The real SCS107 at ~ 2014:187:23:36:36.000 remains.