Upload to lucky on every run of the task
Change logic from just uploading to lucky on changed states (or
in testing) to uploading the h5 file at every run of the tool.
The parsing on the OCC side is just going to grab the file once a
day anyway, so this should make the process more robust to the
occassional failure when pushing the file.